% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{adapt.prior.plot.code}
\alias{adapt.prior.plot.code}
\title{adapt.prior.plot.code- Create Prior Plot for Adaptive Model}
\usage{
adapt.prior.plot.code(trdata, optima.prior, beta.prior)
}
\arguments{
\item{trdata}{Phylogeny and data in treeplyr format}

\item{optima.prior}{Prior mean and scale parameter for optima}

\item{beta.prior}{Prior mean and scale parameter for half-life}
}
\value{
Plots in ggplot2 format
}
\description{
adapt.prior.plot.code- Create Prior Plot for Adaptive Model
}
