% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{adapt.prior.post.plot.code}
\alias{adapt.prior.post.plot.code}
\title{adapt.prior.post.plot.code- Create Prior vs. Posterior Plot for Adaptive Model}
\usage{
adapt.prior.post.plot.code(
  trdata,
  optima.prior,
  beta.prior,
  post,
  optima,
  beta
)
}
\arguments{
\item{trdata}{Phylogeny and data in treeplyr format}

\item{optima.prior}{Prior mean and scale parameter for optima}

\item{beta.prior}{Prior mean and scale parameter for half-life}

\item{post}{Posterior distribution in stanfit class}

\item{optima}{True optima}

\item{beta}{True beta}
}
\value{
Plots in ggplot2 format
}
\description{
adapt.prior.post.plot.code- Create Prior vs. Posterior Plot for Adaptive Model
}
