% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antler.data.R
\docType{data}
\name{antler.data}
\alias{antler.data}
\title{Antler Size Dataset from Tsuboi et al. (2024), Evol. Biol.}
\format{
A csv format dataset
}
\source{
https://doi.org/10.5061/dryad.kh18932dt
}
\usage{
antler.data
}
\description{
Antler Size Dataset from Tsuboi et al. (2024), Evol. Biol.
}
\keyword{datasets}
