% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch-package.R
\docType{package}
\name{blouch-package}
\alias{blouch-package}
\alias{blouch}
\title{The 'blouch' package.}
\description{
Bayesian Linear Ornstein-Uhlenbeck models for Comparative Hypotheses (BLOUCH) fits adaptive models of continuous trait evolution in a Bayesian framework based on categorical or continuous predictors, and incorporates measurement error.
}
\references{
Stan Development Team (2023). RStan: the R interface to Stan. R package version 2.26.22. https://mc-stan.org
}
