% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.direct.adapt.prep.R
\name{blouch.direct.adapt.prep}
\alias{blouch.direct.adapt.prep}
\title{blouch.direct.adapt.prep - setup dat file to run in Blouch's direct effect and adpative model}
\usage{
blouch.direct.adapt.prep(trdata, Y, Y_error, X, X_error, Z_direct, Z_adaptive)
}
\arguments{
\item{trdata}{An object of the class treedata from function treeplyr}

\item{Y}{Vector containing name of column in treedata containing response variable}

\item{Y_error}{Vector containing name of column in treedata containing error of response variable}

\item{X}{Vector containing name(s) of column in treedata containing predictor variable(s)}

\item{X_error}{Vector containing name(s) of column in treedata containing error of predictor variable(s)}

\item{Z_direct}{Vector containing number of direct effect predictor traits}

\item{Z_adaptive}{Vector containing number of adaptive predictor traits}
}
\value{
dat - list file containing objecs setup for Blouch
}
\description{
blouch.direct.adapt.prep - setup dat file to run in Blouch's direct effect and adpative model
}
