% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.direct.prep.R
\name{blouch.direct.prep}
\alias{blouch.direct.prep}
\title{blouch.direct.prep - prep data to run in Stan's direct effect model}
\usage{
blouch.direct.prep(trdata, Y, Y_error, X, X_error, Z_direct)
}
\arguments{
\item{trdata}{An object of the class treedata from function treeplyr}

\item{Y}{Vector containing name of column in treedata containing response variable}

\item{Y_error}{Vector containing name of column in treedata containing error of response variable}

\item{X}{Vector containing name(s) of column in treedata containing predictor variable(s)}

\item{X_error}{Vector containing name(s) of column in treedata containing error of predictor variable(s)}

\item{Z_direct}{Number of direct effect traits}
}
\value{
dat - list file containing objecs setup for Blouch
}
\description{
blouch.direct.prep - prep data to run in Stan's direct effect model
}
