% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{blouch.reg.adapt.prep}
\alias{blouch.reg.adapt.prep}
\title{blouch.reg.adapt.prep - Setup dat file to run multi-optima adaptive model}
\usage{
blouch.reg.adapt.prep(trdata, Y, Y_error, X, X_error, Z_adaptive, reg.column)
}
\arguments{
\item{trdata}{An object of the class treedata from function treeplyr}

\item{Y}{Vector containing name of column in treedata containing response variable}

\item{Y_error}{Vector containing name of column in treedata containing error of response variable}

\item{X}{Vector containing name(s) of column in treedata containing predictor variable(s)}

\item{X_error}{Vector containing name(s) of column in treedata containing error of predictor variable(s)}

\item{Z_adaptive}{Vector containing number of adaptive predictor traits}

\item{reg.column}{Vector containing name of regime column in treedata$dat}
}
\value{
dat - list file containing objecs setup for Blouch
}
\description{
blouch.reg.adapt.prep - Setup dat file to run multi-optima adaptive model
}
