% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.reg.direct.adapt.prep.R
\name{blouch.reg.direct.adapt.prep}
\alias{blouch.reg.direct.adapt.prep}
\title{blouch.reg.direct.adapt.prep - function to setup dat file for Blouch's multi-optima direct effect adaptive model}
\usage{
blouch.reg.direct.adapt.prep(
  trdata,
  Y,
  Y_error,
  X,
  X_error,
  Z_direct,
  Z_adaptive,
  reg.column
)
}
\arguments{
\item{trdata}{An object of the class treedata from function treeplyr}

\item{Y}{Vector containing name of column in treedata containing response variable}

\item{Y_error}{Vector containing name of column in treedata containing error of response variable}

\item{X}{Vector containing name(s) of column in treedata containing predictor variable(s)}

\item{X_error}{Vector containing name(s) of column in treedata containing error of predictor variable(s)}

\item{Z_direct}{Vector containing number of direct effect predictor traits}

\item{Z_adaptive}{Vector containing number of adaptive predictor traits}

\item{reg.column}{Vector containing name of regime column in treedata$dat}
}
\value{
dat - list file containing objecs setup for Blouch
}
\description{
blouch.reg.direct.adapt.prep - function to setup dat file for Blouch's multi-optima direct effect adaptive model
}
