% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{blouch.reg.direct.mlm.prep}
\alias{blouch.reg.direct.mlm.prep}
\title{blouch.reg.direct.mlm.prep - Setup dat file to run Multilevel Multi-Optima direct effect model}
\usage{
blouch.reg.direct.mlm.prep(
  trdata,
  Y,
  Y_error,
  X,
  X_error,
  Z_direct,
  reg.column,
  hl.prior,
  vy.prior,
  optima.prior,
  beta.prior,
  sigma.prior
)
}
\arguments{
\item{trdata}{An object of the class treedata from function treeplyr}

\item{Y}{Vector containing name of column in treedata containing response variable}

\item{Y_error}{Vector containing name of column in treedata containing error of response variable}

\item{X}{Vector containing name(s) of column in treedata containing predictor variable(s)}

\item{X_error}{Vector containing name(s) of column in treedata containing error of predictor variable(s)}

\item{Z_direct}{Vector containing number of direct effect predictor traits}

\item{reg.column}{Vector containing name of regime column in treedata$dat}

\item{hl.prior}{Vector containing log mu and standard error for half-life prior}

\item{vy.prior}{Numeric with prior for scale of Vy}

\item{optima.prior}{Vector containing mu and standard error for optima prior}

\item{beta.prior}{Vector containing mu and standard error for beta prior}

\item{sigma.prior}{Numeric with prior for scale of sigma}
}
\value{
dat - list file containing objecs setup for Blouch
}
\description{
blouch.reg.direct.mlm.prep - Setup dat file to run Multilevel Multi-Optima direct effect model
}
