% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{blouch.reg.mlm.prep}
\alias{blouch.reg.mlm.prep}
\title{blouch.reg.mlm.prep - Setup dat file to run multilevel multi-optima model}
\usage{
blouch.reg.mlm.prep(
  trdata,
  Y,
  Y_error,
  reg.column,
  hl.prior,
  vy.prior,
  optima.prior,
  sigma.prior
)
}
\arguments{
\item{trdata}{An object of the class treedata from function treeplyr}

\item{Y}{Vector containing name of column in treedata containing response variable}

\item{Y_error}{Vector containing name of column in treedata containing error of response variable}

\item{reg.column}{Vector containing name of regime column in treedata$dat}

\item{hl.prior}{Vector containing log mu and standard error for half-life prior}

\item{vy.prior}{Numeric with prior for scale of Vy}

\item{optima.prior}{Vector containing mu and standard error for optima prior}

\item{sigma.prior}{Numeric with prior for scale of sigma}
}
\value{
dat - list file containing objecs setup for Blouch
}
\description{
blouch.reg.mlm.prep - Setup dat file to run multilevel multi-optima model
}
