% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.helpers.R
\name{calc_adaptive_cov_plot}
\alias{calc_adaptive_cov_plot}
\title{calc_adaptive_cov_plot - Calculate covariance matrix for adaptive covariance plot}
\usage{
calc_adaptive_cov_plot(a, sigma2_y, beta, x)
}
\arguments{
\item{a}{Rate parameter for OU model}

\item{sigma2_y}{Variance of Y}

\item{beta}{Slope parameter}

\item{x}{X axis value}
}
\value{
Variance/covariance matrix
}
\description{
calc_adaptive_cov_plot - Calculate covariance matrix for adaptive covariance plot
}
