% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adaptive_dmX.R
\name{calc_adaptive_dmX}
\alias{calc_adaptive_dmX}
\title{calc_adaptive_dmX - calculate adaptive predictor matrix for Blouch}
\usage{
calc_adaptive_dmX(a, T_term, X)
}
\arguments{
\item{a}{Rate parameter for OU model}

\item{T_term}{T_term parameter for phylogeny}

\item{X}{Predictor}
}
\value{
Adaptive predictor matrix
}
\description{
calc_adaptive_dmX - calculate adaptive predictor matrix for Blouch
}
