% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adaptive_V.R
\name{calc_adaptive_V}
\alias{calc_adaptive_V}
\title{calc_adaptive_V - Calculate adaptive variance/covariance matrix}
\usage{
calc_adaptive_V(phy, a, sigma2_y, beta, sigma2_x, Z_adaptive)
}
\arguments{
\item{phy}{phylogney in NEXUS format}

\item{a}{Rate parameter for OU model}

\item{sigma2_y}{Variance of Y}

\item{beta}{slope}

\item{sigma2_x}{Brownian-motion parameterof X}

\item{Z_adaptive}{Number of adaptive predictors}
}
\value{
Variance/covariance matrix
}
\description{
calc_adaptive_V - Calculate adaptive variance/covariance matrix
}
