% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.helpers.R
\name{calc_direct_V}
\alias{calc_direct_V}
\title{calc_direct_V Calculate direct effect model V/CV matrix}
\usage{
calc_direct_V(phy, sigma2_y, a)
}
\arguments{
\item{phy}{Phylogeny in NEXUS format}

\item{sigma2_y}{Variance of Y}

\item{a}{Rate parameter for OU model}
}
\value{
Variance/Covariance marix
}
\description{
calc_direct_V Calculate direct effect model V/CV matrix
}
