% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mixed_dmX.R
\name{calc_mixed_dmX}
\alias{calc_mixed_dmX}
\title{calc_mixed_dmX - calculate mixed direct effect and adaptive predictor matrix for Blouch}
\usage{
calc_mixed_dmX(a, T_term, X, Z_direct, Z_adaptive)
}
\arguments{
\item{a}{Rate parameter for OU model}

\item{T_term}{T_term parameter for phylogeny}

\item{X}{Predictor}

\item{Z_direct}{Number of direct effect predictor traits}

\item{Z_adaptive}{Number of adaptive predictor traits}
}
\value{
Predictor matrix
}
\description{
calc_mixed_dmX - calculate mixed direct effect and adaptive predictor matrix for Blouch
}
