% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.helpers.R
\name{calc_multiadaptive_cov_plot}
\alias{calc_multiadaptive_cov_plot}
\title{calc_multiadaptive_cov_plot - Calcualte covariance matrix for mult-adaptive covariance plot}
\usage{
calc_multiadaptive_cov_plot(a, sigma2_y, beta, x)
}
\arguments{
\item{a}{Rate parameter for OU model}

\item{sigma2_y}{Variance of Y}

\item{beta}{Slope parameter}

\item{x}{X axis value}

\item{Z_adaptive}{Number of adaptive predictors}

\item{n_reg}{Number of regimes}
}
\value{
Variance/covariance matrix
}
\description{
calc_multiadaptive_cov_plot - Calcualte covariance matrix for mult-adaptive covariance plot
}
