% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cervidae.tree.R
\docType{data}
\name{cervidae.tree}
\alias{cervidae.tree}
\title{Cervidae phylogent from Tsuboi et al. (2024), Evol. Biol.}
\format{
A NEXUS format file
}
\source{
https://doi.org/10.5061/dryad.kh18932dt
}
\usage{
cervidae.tree
}
\description{
Cervidae phylogent from Tsuboi et al. (2024), Evol. Biol.
}
\keyword{datasets}
