% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{concat.factor}
\alias{concat.factor}
\title{concat.factor - For internal Blouch use
Thanks to user "snaut" at stackoverflow, http://stackoverflow.com/users/1999873/snaut}
\usage{
concat.factor(...)
}
\arguments{
\item{...}{vector of factors}
}
\value{
factor
}
\description{
concat.factor - For internal Blouch use
Thanks to user "snaut" at stackoverflow, http://stackoverflow.com/users/1999873/snaut
}
