% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{covariance.prior.post.adapt.plot.code}
\alias{covariance.prior.post.adapt.plot.code}
\title{covariance.prior.post.adapt.plot.code- Create Prior vs. Posterior Covariance Plot for Adaptive Model}
\usage{
covariance.prior.post.adapt.plot.code(hl.prior, vy.prior, beta.prior, post)
}
\arguments{
\item{hl.prior}{Prior mean and scale parameter for Half-life}

\item{vy.prior}{Prior scale parameter for Vy}

\item{beta.prior}{Prior mean and scale parameter for beta/slope}

\item{post}{Posterior distribution in stanfit class}
}
\value{
Plots in ggplot2 format
}
\description{
covariance.prior.post.adapt.plot.code- Create Prior vs. Posterior Covariance Plot for Adaptive Model
}
