% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{direct.prior.plot.code}
\alias{direct.prior.plot.code}
\title{direct.prior.plot.code- Create Prior Plot for Direct Effect Model}
\usage{
direct.prior.plot.code(trdata, optima.prior, beta.prior)
}
\arguments{
\item{trdata}{Phylogeny and data in treeplyr format}

\item{optima.prior}{Prior mean and scale parameter for optima}

\item{beta.prior}{Prior mean and scale parameter for half-life}
}
\value{
Plots in ggplot2 format
}
\description{
direct.prior.plot.code- Create Prior Plot for Direct Effect Model
}
