% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{hl.prior.plot.code}
\alias{hl.prior.plot.code}
\title{hl.prior.plot.code- Create Plot for Half-life Prior Distribution}
\usage{
hl.prior.plot.code(hl.prior)
}
\arguments{
\item{hl.prior}{Prior log mean and log sd for half-life}
}
\value{
Plots in ggplot2 format
}
\description{
hl.prior.plot.code- Create Plot for Half-life Prior Distribution
}
