% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{hl.prior.post.emp.plot.code}
\alias{hl.prior.post.emp.plot.code}
\title{hl.prior.post.emp.plot.code- Create Prior vs. Posterior Plot for Half-life Prior Distribution for Empirical Dataset}
\usage{
hl.prior.post.emp.plot.code(hl.prior, post)
}
\arguments{
\item{hl.prior}{Prior log mean and log sd for half-life}

\item{post}{Posterior distribution of stanfit class}
}
\value{
Plots in ggplot2 format
}
\description{
hl.prior.post.emp.plot.code- Create Prior vs. Posterior Plot for Half-life Prior Distribution for Empirical Dataset
}
