% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{hl.prior.post.plot.code}
\alias{hl.prior.post.plot.code}
\title{hl.prior.plot.code- Create Prior vs Posterior Plot for Half-life Prior Distribution for Simulated Dataset}
\usage{
hl.prior.post.plot.code(hl.prior, post, hl)
}
\arguments{
\item{hl.prior}{Prior log mean and log sd for half-life}

\item{post}{Posterior distribution of stanfit class}

\item{hl}{True half-life value}
}
\value{
Plots in ggplot2 format
}
\description{
hl.prior.plot.code- Create Prior vs Posterior Plot for Half-life Prior Distribution for Simulated Dataset
}
