% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{lineage.nodes}
\alias{lineage.nodes}
\title{lineage.nodes - Function for internal Blouch use
Given a certain node, return the list of all parent nodes back to the root of the tree}
\usage{
lineage.nodes(phy, x)
}
\arguments{
\item{phy}{phylogeny in NEXUS format}

\item{x}{node of interest}
}
\value{
list Given a certain node, return the list of all parent nodes back to the root of the tree
}
\description{
lineage.nodes - Function for internal Blouch use
Given a certain node, return the list of all parent nodes back to the root of the tree
}
