% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{optima.prior.plot.code}
\alias{optima.prior.plot.code}
\title{optima.prior.plot.code- Create Prior Plot for Multi-Optima Model}
\usage{
optima.prior.plot.code(trdata, optima.prior)
}
\arguments{
\item{trdata}{Phylogeny and data in treeplyr format}

\item{optima.prior}{Prior mean and scale parameter for optima}
}
\value{
Plots in ggplot2 format
}
\description{
optima.prior.plot.code- Create Prior Plot for Multi-Optima Model
}
