% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{parent}
\alias{parent}
\title{parent function - Returns parent node of offspring node given node number}
\usage{
parent(phy, x)
}
\arguments{
\item{phy}{phylogeny in phytools format}

\item{x}{node number}
}
\value{
value Returns parent node of offspring node given node number
}
\description{
parent function - Returns parent node of offspring node given node number
}
