% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{reg.direct.prior.plot.code}
\alias{reg.direct.prior.plot.code}
\title{reg.direct.prior.plot.code- Create Prior Plot for Multi-Optima Direct Effect Model - Simulated Data}
\usage{
reg.direct.prior.plot.code(trdata, optima.prior, beta.prior)
}
\arguments{
\item{trdata}{Phylogeny and data in treeplyr format}

\item{optima.prior}{Prior mean and scale parameter for optima}

\item{beta.prior}{Prior mean and scale parameter for half-life}
}
\value{
Plots in ggplot2 format
}
\description{
reg.direct.prior.plot.code- Create Prior Plot for Multi-Optima Direct Effect Model - Simulated Data
}
