% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{reg.direct.ve.prior.post.emp.plot.code}
\alias{reg.direct.ve.prior.post.emp.plot.code}
\title{reg.direct.ve.prior.post.emp.plot.code- Create Prior vs. Posterior Plot for Multi-Optima Direct Effect Model - Varying Effects}
\usage{
reg.direct.ve.prior.post.emp.plot.code(
  trdata,
  optima.prior,
  beta.prior,
  post,
  optima,
  beta
)
}
\arguments{
\item{trdata}{Phylogeny and data in treeplyr format}

\item{optima.prior}{Prior mean and scale parameter for optima}

\item{beta.prior}{Prior mean and scale parameter for half-life}

\item{post}{Posterior distribution in stanfit class}
}
\value{
Plots in ggplot2 format
}
\description{
reg.direct.ve.prior.post.emp.plot.code- Create Prior vs. Posterior Plot for Multi-Optima Direct Effect Model - Varying Effects
}
