% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{sig.prior.plot.code}
\alias{sig.prior.plot.code}
\title{sigma.prior.plot.code- Create Prior vs. Posterior Plot for Sigma}
\usage{
sig.prior.plot.code(sigma.prior)
}
\arguments{
\item{sigma.prior}{Prior for mean and sd parameters of sigma}
}
\value{
Plots in ggplot2 format
}
\description{
sigma.prior.plot.code- Create Prior vs. Posterior Plot for Sigma
}
