% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{sigma.prior.post.plot.code}
\alias{sigma.prior.post.plot.code}
\title{sigma.prior.post.plot.code- Create Prior vs. Posterior Plot for Sigma}
\usage{
\method{sigma}{prior.post.plot.code}(sigma.prior, post)
}
\arguments{
\item{sigma.prior}{Prior mean and standard deviation parmateres}

\item{post}{Posterior distribution of stanfit class}
}
\value{
Plots in ggplot2 format
}
\description{
sigma.prior.post.plot.code- Create Prior vs. Posterior Plot for Sigma
}
