% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.X.estimate.R
\name{sigma.X.estimate}
\alias{sigma.X.estimate}
\title{sigma.X.estimate - function for internal Blouch use}
\usage{
\method{sigma}{X.estimate}(phy, ta, predictor, mv.predictor)
}
\arguments{
\item{phy}{phylogeny in phytools format}

\item{ta}{ta paramerter}

\item{predictor}{vector of X values}

\item{mv.predictor}{vector of measurement error for X values}
}
\value{
List with Brownian mean and Sigma2values
}
\description{
sigma.X.estimate - function for internal Blouch use
}
