% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{sim.adaptive.multi.data}
\alias{sim.adaptive.multi.data}
\title{sim.adaptive.multi.data- Simulate Adaptive Data with Multiple X Traits}
\usage{
sim.adaptive.multi.data(phy, N, Z, hl, vy, Sxx, optima, beta)
}
\arguments{
\item{phy}{An object of the class "phylo"}

\item{N}{Number of tips on tree}

\item{Z}{Number of traits}

\item{hl}{True half-life value}

\item{vy}{True Vy value}

\item{Sxx}{Instantaneous variance of the BM process}

\item{optima}{True optima}

\item{beta}{True beta}
}
\value{
Merged phylogeny and data in treeplyr format
}
\description{
sim.adaptive.multi.data- Simulate Adaptive Data with Multiple X Traits
}
