% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{sim.direct.adaptive.data}
\alias{sim.direct.adaptive.data}
\title{sim.adaptive.multi.data- Simulate Direct Effect + Adaptive Data with Multiple X Traits}
\usage{
sim.direct.adaptive.data(
  phy,
  N,
  Z_direct,
  Z_adaptive,
  hl,
  vy,
  Sxx,
  optima,
  beta
)
}
\arguments{
\item{phy}{An object of the class "phylo"}

\item{N}{Number of tips on tree}

\item{hl}{True half-life value}

\item{vy}{True Vy value}

\item{Sxx}{Instantaneous variance of the BM process}

\item{optima}{True optima}

\item{beta}{True beta}

\item{Z}{Number of traits}
}
\value{
Merged phylogeny and data in treeplyr format
}
\description{
sim.adaptive.multi.data- Simulate Direct Effect + Adaptive Data with Multiple X Traits
}
