% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{sim.reg.data}
\alias{sim.reg.data}
\title{sim.reg.data- Simulate Multi-Optima Data}
\usage{
sim.reg.data(phy, N, hl, vy, Sxx, optima, shifts)
}
\arguments{
\item{phy}{An object of the class "phylo"}

\item{N}{Number of tips on tree}

\item{hl}{True half-life value}

\item{vy}{True Vy value}

\item{Sxx}{Instantaneous variance of the BM process}

\item{optima}{True optima}

\item{shifts}{Nodes for regime shifts}
}
\value{
Merged phylogeny and data in treeplyr format
}
\description{
sim.reg.data- Simulate Multi-Optima Data
}
