% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{sim.reg.direct.adapt.data}
\alias{sim.reg.direct.adapt.data}
\title{sim.reg.direct.adapt.data- Simulate Multi-Optima Direct EFfect + Adaptive Model Data}
\usage{
sim.reg.direct.adapt.data(
  phy,
  N,
  Z_direct,
  Z_adaptive,
  hl,
  vy,
  Sxx,
  optima,
  beta,
  shifts
)
}
\arguments{
\item{phy}{An object of the class "phylo"}

\item{N}{Number of tips on tree}

\item{Z_direct}{Number of Direct Effect X traits}

\item{Z_adaptive}{Number of Adaptive X traits}

\item{hl}{True half-life value}

\item{vy}{True Vy value}

\item{Sxx}{Instantaneous variance of the BM process}

\item{optima}{True optima}

\item{beta}{True beta}

\item{shifts}{Nodes for regime shifts}
}
\value{
Merged phylogeny and data in treeplyr format
}
\description{
sim.reg.direct.adapt.data- Simulate Multi-Optima Direct EFfect + Adaptive Model Data
}
