% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{sim.reg.direct.data}
\alias{sim.reg.direct.data}
\title{sim.reg.direct.data- Simulate Multi-Optima Direct Effect Model Data}
\usage{
sim.reg.direct.data(phy, N, Z_direct, hl, vy, Sxx, optima, beta, shifts)
}
\arguments{
\item{phy}{An object of the class "phylo"}

\item{N}{Number of tips on tree}

\item{Z_direct}{Number of X traits}

\item{hl}{True half-life value}

\item{vy}{True Vy value}

\item{Sxx}{Instantaneous variance of the BM process}

\item{optima}{True optima}

\item{beta}{True beta}

\item{shifts}{Nodes for regime shifts}
}
\value{
Merged phylogeny and data in treeplyr format
}
\description{
sim.reg.direct.data- Simulate Multi-Optima Direct Effect Model Data
}
