% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.10K.R
\docType{data}
\name{tree.10K}
\alias{tree.10K}
\title{Primate tree with 301 tip species
This is the primate phylogeny for Version 3 of 10K trees}
\format{
A NEXUS Format dataset
}
\source{
https://10ktrees.nunn-lab.org/
}
\usage{
tree.10K
}
\description{
Primate tree with 301 tip species
This is the primate phylogeny for Version 3 of 10K trees
}
\keyword{datasets}
