% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_fxn.R
\name{ts_fxn}
\alias{ts_fxn}
\title{Internal Blouch function to return tree data}
\usage{
ts_fxn(phy)
}
\arguments{
\item{phy}{phylogeny in phytools format}
}
\value{
A list object with tree data
}
\description{
Internal Blouch function to return tree data
}
