% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{vy.prior.plot.code}
\alias{vy.prior.plot.code}
\title{vy.prior.plot.code- Create plot for Vy Prior Distribution}
\usage{
vy.prior.plot.code(vy.prior)
}
\arguments{
\item{vy.prior}{Prior log mean and log sd for half-life}
}
\value{
Plots in ggplot2 format
}
\description{
vy.prior.plot.code- Create plot for Vy Prior Distribution
}
