% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{vy.prior.post.emp.plot.code}
\alias{vy.prior.post.emp.plot.code}
\title{vy.prior.post.emp.plot.code- Create Prior vs. Posterior Plot for Vy for Empirical Dataset}
\usage{
vy.prior.post.emp.plot.code(vy.prior, post)
}
\arguments{
\item{vy.prior}{Prior log mean and log sd for half-life}

\item{post}{Posterior distribution of stanfit class}
}
\value{
Plots in ggplot2 format
}
\description{
vy.prior.post.emp.plot.code- Create Prior vs. Posterior Plot for Vy for Empirical Dataset
}
