% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plots.R
\name{vy.prior.post.plot.code}
\alias{vy.prior.post.plot.code}
\title{vy.prior.post.plot.code- Create Prior vs. Posterior Plot for Vy}
\usage{
vy.prior.post.plot.code(vy.prior, post, vy)
}
\arguments{
\item{vy.prior}{Prior scale parameter}

\item{post}{Posterior distribution of stanfit class}

\item{hl}{True Vy}
}
\value{
Plots in ggplot2 format
}
\description{
vy.prior.post.plot.code- Create Prior vs. Posterior Plot for Vy
}
