% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{weights_regimes}
\alias{weights_regimes}
\title{weights_regimes - For internal Blouch use
For individual lineage, sum up the segments in each regimes}
\usage{
weights_regimes(a, lineage)
}
\arguments{
\item{a}{Rate parameter from OU model}

\item{lineage}{Individual regime values for lineage}
}
\value{
Return named vector with regimes weights for individual lineage
}
\description{
weights_regimes - For internal Blouch use
For individual lineage, sum up the segments in each regimes
}
