% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch-package.R
\docType{package}
\name{blouch-package}
\alias{blouch-package}
\alias{blouch}
\title{The 'blouch' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2023). RStan: the R interface to Stan. R package version 2.26.22. https://mc-stan.org
}
