% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.helpers.R
\name{calc_adaptive_dmX}
\alias{calc_adaptive_dmX}
\title{calc_adaptive_dmX - calculate adaptive predictor matrix for Blouch}
\usage{
calc_adaptive_dmX(phy, a, X)
}
\arguments{
\item{phy}{Phylogeny in NEXUS format}

\item{a}{Rate parameter for OU model}

\item{X}{Predictor}
}
\value{
Adaptive predictor matrix
}
\description{
calc_adaptive_dmX - calculate adaptive predictor matrix for Blouch
}
