% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.constructor.R
\name{lineage.constructor}
\alias{lineage.constructor}
\title{lineage.constructor function - for internal Blouch use
Construct a list with variables based on regime timing and placement}
\usage{
lineage.constructor(phy, e, anc_maps = "regimes", regimes, ace)
}
\arguments{
\item{phy}{phylogeny in phytools format}

\item{e}{Lineage number}

\item{anc_maps}{Vector with name of regime placement type}

\item{regimes}{Regimes in factor format}

\item{ace}{No real function now}
}
\value{
list with information about individual regime lineages
}
\description{
lineage.constructor function - for internal Blouch use
Construct a list with variables based on regime timing and placement
}
