% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.mlm.data.R
\name{sim.reg.adapt.mlm.ve.data}
\alias{sim.reg.adapt.mlm.ve.data}
\title{sim.reg.adapt.mlm.ve.data- Simulate Multilevel Multi-Optima Adaptive Model Data - Varying Effects}
\usage{
sim.reg.adapt.mlm.ve.data(
  phy,
  N,
  Z_adaptive,
  hl,
  vy,
  Sxx,
  optima.bar,
  optima.sd,
  beta.bar,
  beta.sd,
  rho,
  shifts
)
}
\arguments{
\item{phy}{An object of the class "phylo"}

\item{N}{Number of tips on tree}

\item{Z_adaptive}{Number of Adaptive X traits}

\item{hl}{True half-life value}

\item{vy}{True Vy value}

\item{Sxx}{Instantaneous variance of the BM process}

\item{optima.bar}{True average optima}

\item{optima.sd}{True standard deviation of the optima}

\item{beta.bar}{True average beta}

\item{beta.sd}{True standard deviation of the betas}

\item{rho}{True corelation between optimas and betas}

\item{shifts}{Nodes for regime shifts}
}
\value{
Merged phylogeny and data in treeplyr format
}
\description{
sim.reg.adapt.mlm.ve.data- Simulate Multilevel Multi-Optima Adaptive Model Data - Varying Effects
}
