% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{ts_fxn}
\alias{ts_fxn}
\title{ts_fxn function - internal Blouch function to return tree data}
\usage{
ts_fxn(phy)
}
\arguments{
\item{phy}{Phylogeny in NEXUS format}
}
\value{
A list object with tree data
}
\description{
ts_fxn function - internal Blouch function to return tree data
}
