% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.matrix.R
\name{weight.matrix}
\alias{weight.matrix}
\title{weight.matrix - for internal Blouch use
Wrapper to apply weights_regimes to each lineage}
\usage{
weight.matrix(phy, a, lineages)
}
\arguments{
\item{phy}{phylogeny in phytools format}

\item{a}{OU rate parameter}

\item{lineages}{Vector of regime values}
}
\value{
weights for each lineage
}
\description{
weight.matrix - for internal Blouch use
Wrapper to apply weights_regimes to each lineage
}
