% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouch.prep.R
\name{weights_segments}
\alias{weights_segments}
\title{weights_segments - For internal Blouch use
For individual lineage, determine the weighting of each segment}
\usage{
weights_segments(a, lineage)
}
\arguments{
\item{a}{Rate parameter for OU model}

\item{lineage}{Individual lineage regime values}
}
\value{
For individual lineage, determine the weighting of each segment
}
\description{
weights_segments - For internal Blouch use
For individual lineage, determine the weighting of each segment
}
